% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-minimal.R, R/analyse.R
\name{analyse.minimal_implementation}
\alias{analyse.minimal_implementation}
\alias{analyse}
\title{Analytical solutions for the network}
\usage{
\method{analyse}{minimal_implementation}(implementation, stopping_rule,
  init_signals = "randnorm", init_parameters = "randortho",
  fixed_parameters = FALSE, instantaneous_signals = TRUE,
  tau_estimation, tau_inference, timestep)

analyse(implementation, stopping_rule, init_signals = "randnorm",
  init_parameters = "randortho", fixed_parameters = FALSE,
  instantaneous_signals = TRUE, tau_estimation, tau_inference, timestep)
}
\arguments{
\item{implementation}{An implementation.}

\item{stopping_rule}{A [stopping_rule()]}

\item{init_signals}{What is the signal value at time 0? As a string,
provides the random method to draw the parameters. Currently, only "randnorm"
is implemented.}

\item{init_parameters}{What is the parameter value at time 0? As a string,
provides the random method to draw the parameters. Currently, only
"randortho" is implemented.}

\item{fixed_parameters}{Should the parameters be fixed or can they be
learned?}

\item{instantaneous_signals}{Should the signals be computed instantaneously}

\item{tau_estimation}{What is the time constant of estimation? Is not used if
fixed_parameters is TRUE}

\item{tau_inference}{What is the time constant of inference? Is not used if
instantaneous signals is TRUE}

\item{timestep}{What resolution should the explicit learning
trajectory have?}
}
\description{
This function provides analytical solutions if they are available.
}
\section{Methods (by class)}{
\itemize{
\item \code{minimal_implementation}: At the moment, there are two explicit learning
trajectories available: [explicit_inference()] with fixed parameters (section
C.2) and [explicit_estimation()] with instantaneous signals. This function
provides an interface with these functions that is consistent with the
function [simulate()].
}}

