% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-helper.R
\name{.distance}
\alias{.distance}
\title{Distance helper}
\usage{
.distance(list_1, list_2, type)
}
\arguments{
\item{list_1}{first list}

\item{list_2}{second list}

\item{type}{type of \code{\link[=norm]{norm()}}}
}
\description{
This function computes the distance for a list of matrices.
}
