% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnn-minimal-gd.R
\name{.minimal_fnn_gd_step}
\alias{.minimal_fnn_gd_step}
\title{Computes the gradient descent step for a minimal FNN given a minimal task}
\usage{
.minimal_fnn_gd_step(minimal_task, step_size)
}
\arguments{
\item{minimal_task}{A minimal task}

\item{step_size}{The step size of gradient descent}
}
\description{
Computes the gradient descent step for a minimal FNN given a minimal task
}
