% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit_estimation.R
\name{explicit_estimation}
\alias{explicit_estimation}
\title{Explicit estimation}
\usage{
explicit_estimation(implementation, length_time,
  init_parameters = "randortho", time_constant, timestep)
}
\arguments{
\item{implementation}{an implementation.}

\item{length_time}{The length of the trajectory computation}

\item{init_parameters}{What are the parameters?}

\item{time_constant}{What is the time constant?}

\item{timestep}{What is the resolution of the trajectory?}
}
\description{
This function implements theorem C.1 and returns the explicit estimation.
}
