% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit_inference.R
\name{explicit_inference}
\alias{explicit_inference}
\title{Explicit inference}
\usage{
explicit_inference(implementation, length_time, parameters = "randortho",
  init_signals = "randnorm", time_constant, timestep)
}
\arguments{
\item{implementation}{An implementation of a two-layered linear predictive
coding network}

\item{length_time}{How long should the explicit learning trajectory be
provided?}

\item{parameters}{What are the parameters?}

\item{init_signals}{How are the signals initialized?}

\item{time_constant}{What is the time constant of inference?}

\item{timestep}{What is the resolution of the learning trajectories?}
}
\description{
This function returns the explicit inference as described in section C.2.
}
