% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameters.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Generate parameters}
\usage{
generate_parameters(parameter, implementation)
}
\arguments{
\item{parameter}{provided parameter. Might be a character, e. g. "randortho"}

\item{implementation}{an implementation}
}
\description{
Generates parameters, either by instantiating them or asserting that they are
valid.
}
