% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signals.R
\name{generate_signals}
\alias{generate_signals}
\title{Generate signals}
\usage{
generate_signals(signal, implementation)
}
\arguments{
\item{signal}{provided signal. Might be a character, e. g. "randnorm"}

\item{implementation}{an implementation}
}
\description{
Generates signals, either by instantiating them or asserting that they are
valid.
}
