% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{hpc}
\alias{hpc}
\alias{is_hpc}
\alias{hpc_architecture}
\alias{hpc_input}
\alias{n_samples}
\title{Hierarchical Predictive Coding Model}
\usage{
hpc(architecture, input)

is_hpc(x)

hpc_architecture(hpc)

hpc_input(hpc, time = NULL)

n_samples(hpc)
}
\arguments{
\item{architecture}{an object of class [architecture](is_architecture()).
Lowest layer of the architecture corresponds to the input.}

\item{input}{the input data. Either a function of time, or a matrix.}

\item{x}{an object.}

\item{time}{If specified, returns input at this time as a matrix, where one
column corresponds to one observation}
}
\description{
Hierarchical Predictive Coding Model
}
\section{Functions}{
\itemize{
\item \code{hpc_architecture}: Recovers the architecture

\item \code{hpc_input}: Recovers the input

\item \code{n_samples}: describes how many simples the input provides
}}

\section{Lifecycle}{

The function will gain new arguments that allow specification of other fixed
values than the input. Due to their default values, the standard behavior
will remain the same, however.
}

