% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{implement_minimal}
\alias{implement_minimal}
\alias{implementation_hpc}
\alias{error_weights}
\title{Implementation}
\usage{
implement_minimal(hpc, error_weights = NULL, ...)

implementation_hpc(implementation)

error_weights(implementation)
}
\arguments{
\item{hpc}{Which network should be implemented?}

\item{error_weights}{How should the error within one layer be weighted?
List of matrices. Default is identity matrices in every layer.}

\item{implementation}{An implementation}
}
\description{
This function specifies how the \code{\link[=hpc]{hpc()}} network should be implemented:
should it use Rao & Ballards model or the Free Energy Principle? What
assumptions do we make about the covariance matrix and should it be
learnable?
}
\details{
If you wish to add your own implementation function, you may do so by
providing a class that subsets the general class 'implementation' and may be
called by \code{\link[=simulate]{simulate()}}.
}
\section{Functions}{
\itemize{
\item \code{implementation_hpc}: Recover hpc

\item \code{error_weights}: Recover error weights
}}

\section{Minimal implementation}{

This implementation consists of an optional weight matrix for each layer
}

