% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_signals.R
\name{infer_signals}
\alias{infer_signals}
\title{Infer signals}
\usage{
infer_signals(implementation, parameters, input,
  init_signals = "randnorm", stopping_rule, timestep,
  bypass_assertions = FALSE)
}
\arguments{
\item{implementation}{The implementation}

\item{parameters}{the parameters.}

\item{init_signals}{What is the signal value at time 0? As a string,
provides the random method to draw the parameters. Currently, only
"randortho" is implemented.}

\item{stopping_rule}{The \code{\link[=stopping_rule]{stopping_rule()}}}

\item{timestep}{The timestep}

\item{bypass_assertions}{Should the assertions be skipped to save time? Set
TRUE only if you can trust the input.}
}
\description{
This function returns the inferred signals of a predictive coding
implementation.
}
