% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Extract the input data}
\usage{
input(object, ...)
}
\arguments{
\item{object}{The object to extract the input from, usually a task}

\item{...}{Further arguments provided to the methods}
}
\description{
Extract the input data
}
