% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_analytical.R
\name{is_analytical}
\alias{is_analytical}
\title{Analytical solutions for predictive coding networks}
\usage{
is_analytical(implementation)
}
\description{
This function returns whether an analytical solution exists for this network.
}
