% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{is_architecture}
\alias{is_architecture}
\alias{layered_architecture}
\alias{print.layered_architecture}
\alias{print.linear_layered_architecture}
\title{Architectures}
\usage{
is_architecture(x)

layered_architecture(layers, transformation = identity,
  derivative = function(x) 0 * x + 1)

\method{print}{layered_architecture}(x, ...)

\method{print}{linear_layered_architecture}(x, ...)
}
\arguments{
\item{layers}{an \code{integer} vector. First element provides the input size.
Every other element provides a deeper layer.}

\item{transformation}{What is the nonlinearity? Either a function or a list
of functions with the same length as \code{layers}. Default is the identity
function. Functions must be vectorized.}
}
\description{
Architectures provide the structure of the generative model in a Hierarchical
Predictive Coding network.
}
\section{Functions}{
\itemize{
\item \code{layered_architecture}: defines fully connected layers of size layers
}}

\section{Lifecycle}{

These architectures are going to be extended, but the currently provided
architectures will remain available. The GPU implementation might imply
breaking changes.
}

