% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnn-minimal.R
\name{minimal_fnn}
\alias{minimal_fnn}
\alias{dim.minimal_fnn}
\alias{print.minimal_fnn}
\alias{is_minimal_fnn}
\alias{theta_0}
\alias{theta_1}
\title{Minimal Feedforward Networks}
\usage{
minimal_fnn(theta_0, theta_1)

\method{dim}{minimal_fnn}(x)

\method{print}{minimal_fnn}(x, ...)

is_minimal_fnn(x)

theta_0(minimal_fnn)

theta_1(minimal_fnn)
}
\arguments{
\item{theta_0}{Input the matrix theta_0}

\item{theta_1}{Input the matrix theta_1}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{minimal_fnn}{a minimal FNN}
}
\description{
A minimal feedforward network estimates a q-dimensional vector Y from a
p-dimensional vector X. It is defined by an r x p-matrix theta_0 and a
q x r-matrix theta_1.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: recovers p, q and r
}}

\examples{
minimal_fnn_manual <- minimal_fnn(theta_0 = matrix(-1), theta_1 = matrix(-1))

}
