% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia-minimal.R
\name{minimal_ia}
\alias{minimal_ia}
\title{Minimal integrated account}
\usage{
minimal_ia(lambda = 1, p = NA_integer_, q = NA_integer_,
  r = NA_integer_)
}
\arguments{
\item{lambda}{Interpolation parameter. Can be a vector but may only include
values between zero and one.}

\item{p}{Dimension of $X$. Default is NA.}

\item{q}{Dimension of $Y$. Default is NA.}

\item{r}{Dimension of $H$. Default is NA.}
}
\description{
The minimal integrated account consists of one hidden layer with no prior,
only includes linear transformation, and is concerned with singular tasks.
}
