% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-minimal.R
\name{minimal_task}
\alias{minimal_task}
\alias{weights.minimal_task}
\alias{input.minimal_task}
\alias{output.minimal_task}
\title{Minimal tasks}
\usage{
minimal_task(x, y, w = rep(1, ncol(y)))

\method{weights}{minimal_task}(object, ...)

\method{input}{minimal_task}(object, ...)

\method{output}{minimal_task}(object, ...)
}
\arguments{
\item{x}{Input data}

\item{y}{Output data}

\item{w}{Weights of the loss function}
}
\description{
Minimal tasks
}
