% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Extract the output of the data}
\usage{
output(object, ...)
}
\arguments{
\item{object}{The object to extract the output from, usually a task}

\item{...}{Further arguments provided to the methods}
}
\description{
Extract the output of the data
}
