% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_step.R
\name{signal_step}
\alias{signal_step}
\title{Signal step}
\usage{
signal_step(implementation, parameters, input, signals, time_constant,
  timestep)
}
\arguments{
\item{implementation}{the implementation}

\item{parameters}{the parameters at the particular time}

\item{input}{the input at the particular time}

\item{signals}{the signals at the particular time}

\item{time_constant}{time_constant for the gradient descent (tau_inference)}

\item{timestep}{timestep of the simulation}
}
\description{
This function provides one step for the signal within a predictive coding
implementation
}
