% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameters.R, R/generate_signals.R
\name{simulate-minimal-helpers}
\alias{simulate-minimal-helpers}
\alias{assert_parameters}
\alias{assert_signals}
\title{Helpers for the minimal simulation}
\usage{
assert_parameters(parameters, implementation)

assert_signals(signals, implementation)
}
\description{
Helpers for the minimal simulation
}
\section{Functions}{
\itemize{
\item \code{assert_parameters}: checks whether parameters are valid

\item \code{assert_signals}: checks whether the signals have the
right format
}}

