% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-minimal.R, R/simulation.R
\name{simulate.minimal_implementation}
\alias{simulate.minimal_implementation}
\alias{simulate}
\title{Simulation}
\usage{
\method{simulate}{minimal_implementation}(implementation, stopping_rule,
  tau_estimation, tau_inference, timestep, init_signals = "randnorm",
  init_parameters = "randortho", fixed_parameters = FALSE,
  instantaneous_signals = FALSE,
  instantaneous_signals_stopping_rule = stopping_rule,
  instantaneous_signals_timestep = timestep,
  force_orthogonality = TRUE, ...)

simulate(implementation, stopping_rule, init_signals = "randnorm",
  init_parameters = "randortho", fixed_parameters = FALSE,
  instantaneous_signals = FALSE, tau_estimation, tau_inference, timestep,
  ...)
}
\arguments{
\item{implementation}{The implementation we are concerned with}

\item{stopping_rule}{a [stopping_rule()]}

\item{tau_estimation}{What is the time constant of estimation? Is not used if
fixed_parameters is TRUE}

\item{tau_inference}{What is the time constant of inference? Is not used if
instantaneous signals is TRUE}

\item{timestep}{timestep of the simulation}

\item{init_signals}{What is the signal value at time 0? As a string,
provides the random method to draw the parameters. Currently, only "randnorm"
is implemented.}

\item{init_parameters}{What is the parameter value at time 0? As a string,
provides the random method to draw the parameters. Currently, only "randortho"
is implemented.}

\item{fixed_parameters}{Should the parameters be fixed or can they be
learned?}

\item{instantaneous_signals}{Should the signals be computed instantaneously,
i. e. should we either use a closed form for the signals or use the EM
algorithm?}

\item{force_orthogonality}{Should parameters be orthogonalized?}
}
\description{
This function simulates a particular implementation of a predictive coding
network.
}
\section{Methods (by class)}{
\itemize{
\item \code{minimal_implementation}: This simulates the minimal implementation
}}

