% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformations}
\alias{transformations}
\title{Transformations}
\usage{
transformations(x)
}
\arguments{
\item{x}{an object}
}
\description{
Recovers the transformation in an architecture, network, etc.
}
